pyproject.toml
[project]
name = "nl-sql-cli"
version = "0.1.0"
description = "A CLI tool that converts natural language to SQL queries against the Chinook database"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.96.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
    "typer>=0.24.2",
]

[project.optional-dependencies]
eval = [
    "openai>=1.0.0",
]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
nlsql = "src.cli:app"
